🗜️ squash schemas
