--- src/codeinterpreterapi/schema/__init__.py
@@ -1,13 +1,0 @@
-from .file import File
-from .input import CodeInput, FileInput
-from .response import CodeInterpreterResponse, UserRequest
-from .status import SessionStatus
-
-__all__ = [
-    "CodeInterpreterResponse",
-    "CodeInput",
-    "File",
-    "FileInput",
-    "UserRequest",
-    "SessionStatus",
-]
