--- src/codeinterpreterapi/schema/status.py
@@ -1,10 +1,0 @@
-from codeboxapi.schema import CodeBoxStatus  # type: ignore
-
-
-class SessionStatus(CodeBoxStatus):
-    @classmethod
-    def from_codebox_status(cls, cbs: CodeBoxStatus) -> "SessionStatus":
-        return cls(status=cbs.status)
-
-    def __repr__(self) -> str:
-        return f"<SessionStatus status={self.status}>"
